src/test/java/ru/coffee/lettercount/controller/LetterControllerTest.java
package ru.coffee.lettercount.controller;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import ru.coffee.lettercount.util.LetterCounter;

import java.util.LinkedHashMap;
import java.util.Map;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.anyString;
import static org.mockito.Mockito.doReturn;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class LetterControllerTest {

    @Mock
    private LetterCounter letterCounter;

    @InjectMocks
    private LetterController letterController;

    @Test
    public void validResponseEntity() {

        LinkedHashMap<Character, Long> linkedHashMap = new LinkedHashMap<>();
        ResponseEntity<?> responseEntity = letterController.countLetter(anyString());

        assertNotNull(responseEntity);
        assertEquals(HttpStatus.OK, responseEntity.getStatusCode());
        assertEquals(MediaType.APPLICATION_JSON, responseEntity.getHeaders().getContentType());
        assertEquals(linkedHashMap, responseEntity.getBody());
    }


}